moved serializers and viewsets for asset agreements to domain app
